I found the goddamn error with the layout!
The error was that they action of AltGr on the dead keys was swapped, pressing the tilde key put the dead variant, using Shift put another variant instead of ~,
the issue was that i was using "with dead keys" instead of "with AltGr dead keys" ... ... ... Yah, that shit took me months to figure out.

I also set the console.useXkbConfig to true to make my life easier.